+++ Progetto1/pacchettoAcknowledgement.h
@@ -11,7 +11,7 @@ class pacchettoAcknowledgement
 public:
 	pacchettoAcknowledgement(void);
 	// funzione che restituisce la dimensione (ideale, non quella dovuta agli allineamenti 
-	// fatti dal compilatore) in byte del messaggio tenendo anche in conto l'eventuale padding
+	// fatti dal compilatore) in unsigned int del messaggio tenendo anche in conto l'eventuale padding
 	// questa funzione sarà chiamata da chi vorrà serializzare il messaggio, per poter allocare il buffer
 	int getSize(){return 54;};
 	// funzioni di interfaccia set e get per ogni campo dati del pacchetto
@@ -26,9 +26,9 @@ public:
 	void setT_TRAIN(int N){data.T_TRAIN = N;};
 	int getT_TRAIN(){return data.T_TRAIN;};
 	// metodi per la serializzazione e deserializzazione del messaggio
-	// il buffer di byte deve essere stato precedentemente correttamente allocato.
-	void serialize(byte *buff);
-	void deserialize(byte *buff);
+	// il buffer di unsigned int deve essere stato precedentemente correttamente allocato.
+	void serialize(unsigned int *buff);
+	void deserialize(unsigned int *buff);
 	~pacchettoAcknowledgement(void);
 
 	System::String ^ ToString(){
